refactor(zb_cli): change flat structure to delagated modules
