+++ zb_cli/src/cli.rs
@@ -1,0 +1,52 @@
+use clap::{Parser, Subcommand};
+use std::path::PathBuf;
+
+#[derive(Parser)]
+#[command(name = "zb")]
+#[command(about = "Zerobrew - A fast Homebrew-compatible package installer")]
+#[command(version)]
+pub struct Cli {
+    #[arg(long, env = "ZEROBREW_ROOT")]
+    pub root: Option<PathBuf>,
+
+    #[arg(long, env = "ZEROBREW_PREFIX")]
+    pub prefix: Option<PathBuf>,
+
+    #[arg(long, default_value = "48")]
+    pub concurrency: usize,
+
+    #[command(subcommand)]
+    pub command: Commands,
+}
+
+#[derive(Subcommand)]
+pub enum Commands {
+    Install {
+        formula: String,
+        #[arg(long)]
+        no_link: bool,
+    },
+    Uninstall {
+        formula: Option<String>,
+    },
+    Migrate {
+        #[arg(long, short = 'y')]
+        yes: bool,
+        #[arg(long)]
+        force: bool,
+    },
+    List,
+    Info {
+        formula: String,
+    },
+    Gc,
+    Reset {
+        #[arg(long, short = 'y')]
+        yes: bool,
+    },
+    Init,
+    Completion {
+        #[arg(value_enum)]
+        shell: clap_complete::shells::Shell,
+    },
+}
