+++ zb_cli/src/commands/mod.rs
@@ -1,0 +1,9 @@
+pub mod completion;
+pub mod gc;
+pub mod info;
+pub mod init;
+pub mod install;
+pub mod list;
+pub mod migrate;
+pub mod reset;
+pub mod uninstall;
